+++ src/main/java/com/zoi4erom/strategygame/repository/StatisticRepository.java
@@ -3,6 +3,6 @@ package com.zoi4erom.strategygame.repository;
 import com.zoi4erom.strategygame.entity.Statistic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatisticRepository extends JpaRepository<Statistic, Integer> {
+public interface StatisticRepository extends JpaRepository<Statistic, Long> {
 
 }
