+++ src/main/java/com/zoi4erom/strategygame/repository/UnitRepository.java
@@ -4,6 +4,6 @@ import com.zoi4erom.strategygame.entity.Unit;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitRepository extends JpaRepository<Unit, Integer> {
+public interface UnitRepository extends JpaRepository<Unit, Long> {
 	Optional<Unit> findUnitByName(String unitName);
 }
