+++ src/main/java/com/zoi4erom/strategygame/repository/UnitTypeRepository.java
@@ -4,6 +4,6 @@ import com.zoi4erom.strategygame.entity.UnitType;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitTypeRepository extends JpaRepository<UnitType, Integer> {
+public interface UnitTypeRepository extends JpaRepository<UnitType, Long> {
 	Optional<UnitType> findUnitTypeByName(String unitTypeName);
 }
